Moved the practice tests to the file test.py #49
